Create comment routes and update user routes
Create new comment routes to create a comment, update a comment and
delete a comment. Create a new custom errors file for comments which
checks if the author (not the 'owner') of the comment is the current
user trying to make changes. Reviewed user routes and added a line
to pass the username field to the database.